Move CSV logic into a helpers file
