+++ src/components/Books/index.jsx
@@ -1,13 +1,4 @@
-function createCsvItem(book) {
-  return Object.values(book).map(JSON.stringify).join(",")
-}
-
-function jsonToCsv(books = []) {
-  const header = Object.keys(books[0]).join(",")
-  const items = books.map(createCsvItem).join("\n")
-
-  return `${header}\n${items}`
-}
+import { jsonToCsv } from "../../helpers/csv"
 
 export default function Books({ books = [] }) {
   const blob = new Blob([jsonToCsv(books)], { type: "text/csv" })
