+++ src/helpers/csv.jsx
@@ -1,0 +1,10 @@
+function createCsvItem(book) {
+  return Object.values(book).map(JSON.stringify).join(",")
+}
+
+export function jsonToCsv(books = []) {
+  const header = Object.keys(books[0]).join(",")
+  const items = books.map(createCsvItem).join("\n")
+
+  return `${header}\n${items}`
+}
